Fixed a race condition bug that caused corrupted output when running on MacOS using MoltenVK
Found a bug that intermittently caused some of the output values of the FFT kernel to be corrupted. This effect was only reproduced on a Macbook Pro with a M2 Pro GPU that used MoltenVK as it's Vulkan driver. The issue seems to only occur in the FFT stage that handles non-strided data (presumably because this reduced memory latency, which increased the likelihood of the race condition causing an error).

Adding a "memoryBarrier();" call right before each "barrier();" call seems to completely fix the issue.